+++ src/views/UserEdit/UserForm.js
@@ -400,6 +400,7 @@ class UserForm extends React.Component {
                     accordionId="contactInfo"
                     addressTypes={formData.addressTypes}
                     preferredContactTypeId={initialValues.preferredContactTypeId}
+                    preferredEmailCommunication={initialValues.preferredEmailCommunication}
                     disabled={isShadowUser}
                   />
                   <EditCustomFieldsRecord